feat(fri-prover): In witness - panic if protocol version is not available (#192)
# What ❔

* When running witness in continuous mode, panic when you cannot load
the matching protocol version.

## Why ❔

* Witness is running with a given commitment key. At startup, it checks
if there is a matching protocol version for this key, and later uses
this protocol version to fetch its task.
* We are running witness in 2 modes: either 'batch' (where it runs up to
X tasks and quits) or continuous. Batch is mostly used in production,
while continuous is used in local deployments.
* And if in the local deployment, the database was not updated - and
protocol version is missing - then the job would get stuck forever (as
it will try to pickup job for non-existing protocol). So instead, let's
panic explicitly.