feat(vm): Move all vm versions to the one crate (#249)
# What ❔
Move all vm versions to the one crate. 
p.s. You will see a lot of `multivm::vm_latest` in the public interface.
It's temporary.
In the next PR it will be a common interface and we won't reference to
the concrete vm

## Why ❔

In the future, i'm going to unify all vms by one interface and this
interface will be used for Multivm and for one Vm. There will be no more
difference between them. This pr is a preparation step. We need it for
mitigating circular dependencies in the future

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.

Signed-off-by: Danil <deniallugo@gmail.com>