feat(vm): Use the one interface for all vms (#277)
What ❔
Implement one common interface for all vms. It's the first attempt to do
it and it's done only for two latest vms.
In the future all vms will be migrated to the one interface.

Why ❔
We need to unify our work with VM now it's always unpredicable what are
we calling and when


p.s. I'm not going to merge it before boojum 

The thing, i don't like: 
Using `vm_latest::HistoryEenable` in interface. 
It suppose to take another Eternity to refactor HistoryMode 

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.

---------

Signed-off-by: Danil <deniallugo@gmail.com>