fix(api): `debug_trace*` no longer throws error if the `TracerConfig` object is incomplete (#468)
## What ❔

* Derived `Default` for `CallTracerConfig`
* Allow default values for `tracer_config` field inside `TracerConfig` 
* Integration tests for this exact issue to confirm it is solved.
* Running `zk test i api` will now also run the test suite for the debug
namespace. (this is optional and not required in any way for the bug
fix, so I can remove it)

## Why ❔

Check out the discussion in #451 
The following `debug` namespace rpc methods would throw an error if you
would include a `TracerConfig` as an argument but only specified the
`tracer` and not also the `tracer_config` (see issue mentioned above for
full description of the issue):

* `debug_traceBlockByNumber`
*  `debug_traceBlockByHash`
*  `debug_traceCall`
*  `debug_traceTransaction`

Some ETH nodes I tested do not throw an error in this case, so the
zkSync node should behave the same way after this fix.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.