fix(house_keeper): Emit the correct circuit_id for aggregation round 2 (#547)
## What ❔

This is us hardcoding the circuit_id to 2 for aggregation_round 2. This
fix is useless as soon as we change the config (not expected anytime
soon).

A real fix will follow where we address what is saved in database layer
and this patch will be removed altogether.

## Why ❔

This is necessary to enable autoscaler to work (emits the correct data
for prover groups, which in turn can be picked by autoscaler).

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.