chore(ci): Pre-download compilers, as a workaround for old hardhat plugins (#645)
## What ❔

Add hack way to pre-download needed compileres

## Why ❔

Hardhat plugins, which we are using right now, are pretty often rate
limited by github for downloading compilers, due to unoptimized download
procedure. This issue was fixed in later versions of plugins, but we
can't update _some_ of the hardhat plugins due to changes in deps for
them. This pre-download will help us till the time we can update them.

## Checklist


- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `cargo spellcheck
--cfg=./spellcheck/era.cfg --code 1`.