fix(prover): increase DB polling interval for witness vector generators (#697)
Prover DB is under heavy load. 

The query that witness vectors generators run to get a new prover job is
slow. In theory we shouldn't be running any witness vectors gens for
circuits that don't have any jobs queued, but query insights suggest
otherwise.

This increases the polling interval - so that even if there are no jobs,
we don't perform the query too often