refactor(object_store): Wrap ObjectStore in Arc instead of Box (#820)
## What ❔

Changes `Box<dyn ObjectStore>` to `Arc<dyn ObjectStore>`.

## Why ❔

`ObjectStore` is meant to be a point of customization in ZK Stack, but
with the current approach the only way to get it is via
`ObjectStoreFactory`, and it yields objects packed in `Box`, e.g. each
instance is unique.
In order to make it customizable, `ObjectStore` should be universally
shareable (to allow it to be created once and then copied), and this is
the first step toward that.
Later `ObjectStoreFactory` type will be removed completely to grant full
control regarding the initial type to the initialization code.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `cargo spellcheck
--cfg=./spellcheck/era.cfg --code 1`.