feat(vm): Add batch input abstraction (#817)
## What ❔

The new fee model will allow us to have separate L1 gas price and
pubdata price (and it is needed for EIP4844). That's why the
`BatchFeeInput` struct is needed as it provides the needed fields for
both the VM version.

The version that is provided in this PR will require not additional
changes on the server side (i.e. whatever config params that we have
right now is enough, the full integration of the config params for the
new fee model will be added to the main PR:
https://github.com/matter-labs/zksync-era/pull/791/.

While the defined enum allows for both l1-gas-price pegged and
independent pubdata pricing, in this PR only L1Pegged is ever
instantiated

❗ A new JSON-RPC method has been added: `zks_getMainNodeFeeParams`. It
will be used by the EN and it will serve the same purpose as the
`zks_getL1GasPrice`, i.e. retrieving the fee info from the main node,
but now it will be extended to provide the fair l2 gas price as well as
other config params that for now will be defined on the main node. The
`zks_getL1GasPrice` will continue working as usual to not interrupt the
existing external nodes, but may be removed in the long run.

## Why ❔

Splitting the integration of the 1.4.1 & the new fee model
(https://github.com/matter-labs/zksync-era/pull/791/) into separate
smaller PRs

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `cargo spellcheck
--cfg=./spellcheck/era.cfg --code 1`.