feat: made external node tests use a snapshot of the main node db (#913)
## What ❔

Made external node tests use a snapshot of the main node db.
It is implemented by postgres dump & restore in "zk db setup" call.

## Why ❔

Current implementation of the gossip network-based block syncing for
external node needs to know the certificate on the consensus genesis
block (which might not be the first miniblock in the db) to be able to
verify all subsequent blocks.
This certificate cannot be generated by the external node, because it
contains signature of the validator (aka main node). For simplicity, we
assume that this certificate on genesis block is already present in the
external node's db, which has been restored from a snapshot of the main
node db.

Integration tests for external node start with an empty db instead,
which disables fetching blocks over gossip network altogether. This PR
is supposed to fix that.