fix(prover): QoL socket utilization (#1020)
## What ❔

This PR adds necessary timeout for WVG <-> prover socket communication.
On top of it, also improves the debugging experience by adding logs on
what connections are being established to what IPs.

## Why ❔

Without the timeout, we could have WVGs stuck for as long as the
connection takes. In practice, we've observed over 30 minutes.

This will limit such instances to ~1 minute (10 retries * 6 seconds
timeout). Whilst this is a duct-tape, it will do until we get around
implementing the long-term solution.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.