feat: adds debug_traceBlockByNumber.callFlatTracer (#1413)
## What ❔

This PR adds support for ‘Parity style’ output for
`debug_traceBlockByNumber` tracer.
Specifically adds `.callFlatTracer` to modify the format of the tracer
output so that it’s the same format as Parity’s tracer output.

## Why ❔

We want to provide a way to get the the same format as parity in
`debug_traceBlockByNumber`

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [X] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [X] Linkcheck has been run via `zk linkcheck`.