refactor: Switch core workspace to using workspace deps (#1485)
## What ❔

Switches core workspace to using workspace deps.

⚠️ No-nitpick-territory ⚠️ 
During review, please focus on the fact that this should be an
incremental improvement. This PR was _**painful**_ to prepare, and I
would like to merge it ASAP, since it's also likely to get merge
conflicts. If there are non-critical comments, I'll try to address them
in follow-up PRs. I've touched a lot of code that was in imperfect state
before, and now it's in the diff (for example, incorrectly set features
for crates), and maybe some things I've done aren't as pretty as they
can be -- but IMHO the switch to using workspace deps has enough value
to merge it without addressing all and every minor thing beforehand.

Feel free to leave as many comments as you have to, but it'd be great if
they will still come with a ✅ unless I really broke something here.

## Why ❔

- Soft prerequisite for publishing on crates.io.
- More convenient.
- More modern.
- No need to worry that some dependency has multiple versions across
different crates.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.