fix(api): Fix API server shutdown flow (#1425)
## What ❔

- API server now doesn't (gracefully) shut down immediately after
receiving a signal. Instead, it waits for some time (until the inbound
traffic reaches zero), and then starts the graceful shutdown.
- Some tasks run by the main node are now shutdown-aware.

## Why ❔

If the server starts shutting down immediately after receiving a signal
(as it does now), this means that it will drop any new traffic, which is
not what the K8s load balancer expects.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.