feat(prover): export prover traces through OTLP (#1427)
## What ❔

This PR adds an optional `tracing` layer to `vlog` that exports all
spans to a configurable OTEL collector.

This layer is then used in most prover components to export spans with
structured information about the block number that is currently being
processed.

## Why ❔

Better observability for prover and tracing for logs with a specific
block number

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.