feat(genesis): Using genesis config only during the genesis (#1423)
## What ❔

During the genesis, we have some fixed params, and they are changing
only during protocol upgrade. In this PR, I've not created an additional
file that represents genesis state, but only loaded it from
environmental variables.



## Why ❔

For system config refactoring we have to split our config system to a
few different files and it's the first step to it

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.

---------

Signed-off-by: Danil <deniallugo@gmail.com>