feat(en): Make snapshot syncing future-proof (#1441)
## What ❔

Adds a snapshot version to snapshot metadata returned by the
corresponding endpoint. If the version of the returned snapshot is not
supported by the EN, it immediately errors with an informative message.

## Why ❔

We plan to change the snapshot format soon-ish, but it could happen
after the initial public EN release.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.