fix(prover): Remove redundant LoadingMode (#1496)
This is PR 1 of many to move us to safe Prover Protocol Upgrades.

`ProtocolVersionLoadingMode` is a crutch that was added as part of
transitioning to Boojum. This served as a manual intervention mechanism
during protocol upgrades. It was a suboptimal choice (read as technical
debt) done due to rushed migration to boojum.

A proper fix involves reading the information from database, which in
turn is updated by information received from contracts. This change was
done recently (1.4.2 migration), but the old cludge was left in place
(you guessed it, due to rushed migration).

This PR removes this cludge and makes sure that nobody will try to use
it assuming it will work (whilst it does work, it requires "specialized"
knowledge on how to use it and may cause us submitting broken proofs to
L1).