feat(en): Rework storing and using protective reads (#1515)
## What ❔

- Does not persist protective reads if the node opts out of this.
- Handles absent protective reads in Metadata calculator (do not use
them in the lightweight mode; warn in the full mode).

## Why ❔

Protective reads are only used by the full-mode tree to generate input
for the prover. (Technically, they are also used to filter out
inconsequential writes from storage slots touched in an L1 batch, but
this logic can be removed / rewritten.)

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.