feat(configs): Implement new format of configs and implement protobuf for it    (#1501)
## What ❔

Loading new format of config from files. 
Only main node is fully moved to the ability to work from files.
Init process is intentionally has not be adapted for the file-based
configurtaion
This format will be used with the new configuration 


## Why ❔

For having an opportunity not to rely on envs we can load files with
necessary configuration.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.

---------

Signed-off-by: Danil <deniallugo@gmail.com>
Co-authored-by: Grzegorz Prusak <pompon.pompon@gmail.com>