chore: Migrate more components including call tracer to u64 (#1556)
## What ❔

Migrate more parts to u64 in preparation to the v1.5.0.

Important note:
- While the internal `gas` used by the calls in the tracer could remain
u32, the top level call should continue being equal to `gasLimit`
requiring it to be moved to u64
- In our DB call traces are stored as `Vec<u8>` encoded by `bincode`, in
other words no data about the structure is preserved and so we'll have
to decode the trace based on the protocol version

## Why ❔

In preparation for v1.5.0, to reduce the diff on the v1.5.0 PR.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.