test: Fix easy to fix integration tests (#1590)
## What ❔

Fixes a couple of errors in integration tests with known causes.

- Errors in `getFilterChanges` for pending transactions are caused by
introducing the mempool cache. The test makes too strict assumptions as
to the transaction latency, which can be easily relaxed.
- The revert test somewhat frequently fails because an L1 transaction
response for the deposit is null. The exact cause is unknown (maybe,
switching to reth?), but it seems to be fixable by retrying the
transaction retrieval until it's not null.

## Why ❔

Failing integration tests slow down development.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.