fix(vm): Fix storage oracle and estimation (#1634)
## What ❔

- Makes "free" slots also free with regard to I/O costs. These are
context / operator balance slots which are global and warm for everyone.
- Fixes a bug with setting the initial value inside storage oracle. This
bug only affects 1.5.0.
- The gas used for pubdata is also multiplied by the
`estimated_fee_scale_factor`. For "small" transactions, e.g. transfers
this gives us the same estimation overhead as in the previous (v1.4.1)
system. But this will produce a bit larger overhead for big transactions
that involve a lot of pubdata, but this is the correct consistent
behavior and it should not affect most of the users.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.