docs: fixed bytecode compression example (#1013)
## What ❔
Updated the bytecode compression explanation example, to be valid and
working.

In the connecting contracts PR:
https://github.com/matter-labs/era-contracts/pull/193 I added a test
that checks the bytecode and runs the `publishCompressBytecode` function
with the bytecode from this example.

## Why ❔
The previous example was not valid, it could confuse newcomers, because
the example had different bytecode compared to what was described.

## Checklist
- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.