feat: tee_verifier_input_producer (#1860)
## What ❔

Extract all data needed to re-execute and verify an L1Batch without
accessing the DB and/or the object store.

For testing purposes, the L1 batch is re-executed immediately for now.
Eventually, this component will only extract the inputs and send them to
another machine over a "to be defined" channel, e.g., save them to an
object store.

## Why ❔

Have an additional proof 2F verification running inside a TEE

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.

---------

Signed-off-by: Harald Hoyer <harald@matterlabs.dev>
Co-authored-by: Thomas Knauth <tk@matterlabs.dev>
Co-authored-by: Patrick Bęza <pab@matterlabs.dev>