refactor: Deprecated eth_block, eth_hash and deadline_block in Transaction (#2091)
These fields are unused and don't affect the execution. eth_hash and
deadline_block are not even stored, so they have been removed
completely. eth_block is stored in db and is part of the block signed by
consensus, therefore removing it will require a protocol upgrade and a
hard fork.