feat(en): Allow recovery from specific snapshot (#2137)
## What ❔

Allows recovering a node from a specific snapshot specified at the start
of recovery.

## Why ❔

Useful at least for testing recovery and pruning end-to-end on the
testnet. There, L1 batches are produced very slowly, so it makes sense
to recover from an earlier snapshot in order to meaningfully test
pruning.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.